requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
plotly
pyjwt
cryptography
requests
google-cloud-bigquery
db-dtypes
